#191 - Header classes moved to com.artipie.http.headers package
