+++ src/main/java/com/artipie/http/rs/RsWithBody.java
@@ -28,6 +28,7 @@ import com.artipie.asto.Content;
 import com.artipie.http.Connection;
 import com.artipie.http.Headers;
 import com.artipie.http.Response;
+import com.artipie.http.headers.ContentLength;
 import io.reactivex.Flowable;
 import java.nio.ByteBuffer;
 import java.nio.charset.Charset;
