+++ src/test/java/com/artipie/http/headers/AuthorizationBasicTest.java
@@ -21,47 +21,24 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.artipie.http.rs;
+package com.artipie.http.headers;
 
-import com.artipie.http.Headers;
-import com.artipie.http.rq.RqHeaders;
+import org.hamcrest.MatcherAssert;
+import org.hamcrest.core.IsEqual;
+import org.junit.jupiter.api.Test;
 
 /**
- * Content-Length header.
+ * Tests for {@link Authorization.Basic}.
  *
- * @since 0.10
+ * @since 0.12
  */
-public final class ContentLength extends Header.Wrap {
+public final class AuthorizationBasicTest {
 
-    /**
-     * Header name.
-     */
-    public static final String NAME = "Content-Length";
-
-    /**
-     * Ctor.
-     *
-     * @param value Header value.
-     */
-    public ContentLength(final String value) {
-        super(new Header(ContentLength.NAME, value));
-    }
-
-    /**
-     * Ctor.
-     *
-     * @param headers Headers to extract header from.
-     */
-    public ContentLength(final Headers headers) {
-        this(new RqHeaders.Single(headers, ContentLength.NAME).asString());
-    }
-
-    /**
-     * Read header as long value.
-     *
-     * @return Header value.
-     */
-    public long longValue() {
-        return Long.parseLong(this.getValue());
+    @Test
+    void shouldHaveExpectedValue() {
+        MatcherAssert.assertThat(
+            new Authorization.Basic("Aladdin", "open sesame").getValue(),
+            new IsEqual<>("Basic QWxhZGRpbjpvcGVuIHNlc2FtZQ==")
+        );
     }
 }
