+++ src/test/java/com/artipie/http/headers/package-info.java
@@ -21,24 +21,11 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.artipie.http.rs;
-
-import org.hamcrest.MatcherAssert;
-import org.hamcrest.core.IsEqual;
-import org.junit.jupiter.api.Test;
 
 /**
- * Tests for {@link Authorization.Basic}.
+ * Tests for HTTP header classes.
  *
- * @since 0.12
+ * @since 0.13
  */
-public final class AuthorizationBasicTest {
+package com.artipie.http.headers;
 
-    @Test
-    void shouldHaveExpectedValue() {
-        MatcherAssert.assertThat(
-            new Authorization.Basic("Aladdin", "open sesame").getValue(),
-            new IsEqual<>("Basic QWxhZGRpbjpvcGVuIHNlc2FtZQ==")
-        );
-    }
-}
